added shop to Devon Corp, upgraded price value from u16 to u32

+++ include/item.h
@@ -11,7 +11,7 @@ struct Item
 {
     u8 name[ITEM_NAME_LENGTH];
     u16 itemId;
-    u16 price;
+    u32 price;
     u8 holdEffect;
     u8 holdEffectParam;
     const u8 *description;
@@ -65,7 +65,7 @@ u16 CountTotalItemQuantityInBag(u16 itemId);
 bool8 AddPyramidBagItem(u16 itemId, u16 count);
 bool8 RemovePyramidBagItem(u16 itemId, u16 count);
 const u8 *GetItemName(u16 itemId);
-u16 GetItemPrice(u16 itemId);
+u32 GetItemPrice(u16 itemId);
 u8 GetItemHoldEffect(u16 itemId);
 u8 GetItemHoldEffectParam(u16 itemId);
 const u8 *GetItemDescription(u16 itemId);
